agentex/pyproject.toml
[tool.poetry]
name = "agentex"
version = "0.1.0"
description = ""
authors = ["Felix Su <felixs8696@gmail.com>"]
readme = "README.md"
packages = [{ include = "agentex" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
litellm = "^1.48.2"
python-dotenv = "^1.0.1"
temporalio = "^1.7.1"
uvicorn = "^0.30.6"
redis = "^5.1.0"
sqlalchemy = "^2.0.35"
asyncpg = "^0.29.0"
alembic = "^1.13.3"
psycopg2 = "^2.9.9"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
